Move spatial_graph into graph_libs
