+++ src/geff/graph_libs/_spatial_graph.py
